+++ changelog.md
@@ -18,6 +18,7 @@ __[changes]__
 __[maintenance]__
 
 * Update certificate information for CCleaner installers.
+* Update certificate information for Eclipse Temurin JDK 8 LTS installers.
 * Update certificate information for Eclipse Temurin JDK 17 LTS installers.
 * Update certificate information for Eclipse Temurin JRE 17 LTS installers.
 * Update certificate information for Eclipse Temurin JDK 21 LTS installers.